test(wrap): gate peer creds test on linux cgo
